services/judge/requirements.txt
kafka-python
numpy
python-dotenv
river
prometheus-client
psycopg2-binary
